requirements.txt
Babel==2.6.0
Django==2.0.6
Pillow==5.1.0
PyJWT==1.6.4
South==1.0.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.2.5
command-not-found==0.3
defusedxml==0.5.0
django-formtools==2.1
django-haystack==2.8.1
django-localflavor==2.0
django-phonenumber-field==2.0.0
elasticsearch==6.2.0
html5lib==0.999
idna==2.6
language-selector==0.1
oauthlib==2.1.0
phonenumberslite==8.9.8
pycurl==7.19.3
pygobject==3.12.0
python-apt===0.9.3.5ubuntu2
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
ufw===0.34-rc-0ubuntu2
unattended-upgrades==0.1
urllib3==1.22
wheel==0.24.0
~                       
